V-F-1_Lebensbereiche: Show list of existing 'Lebensbereiche'
A RecyclerView is display which contains items which contain the description and the color label of a 'Lebensbereich'. For that a data class has been created and an adapter has been prepared for the Recycler View. Currently a dummmy list from the AreasViewModel.kt is displayed. 10 different labels have been created as drawable shapes.