V-F-3_Sprintplanung: Simplify data structure
Foreign key from area to sprint has been removed as the capacity values are only needed during sprint planning and do not need to be persisted on the db connected to the sprint. In addition, Task has been made parcelable to pass it as an argument in navigation.

Refers to V-F-3.1